+++ packages/react-ui/src/index.d.ts
@@ -5,6 +5,7 @@ import FormulaWidgetUI from './widgets/FormulaWidgetUI';
 import HistogramWidgetUI from './widgets/HistogramWidgetUI';
 import PieWidgetUI from './widgets/PieWidgetUI';
 import LegendWidgetUI, { LEGEND_TYPES } from './legend/LegendWidgetUI';
+import ScatterPlotWidgetUI from './widgets/ScatterPlotWidgetUI';
 
 export {
   cartoThemeOptions,
@@ -13,6 +14,7 @@ export {
   FormulaWidgetUI,
   HistogramWidgetUI,
   PieWidgetUI,
+  ScatterPlotWidgetUI,
   LegendWidgetUI,
   LEGEND_TYPES,
 };