+++ packages/react-widgets/src/index.d.ts
@@ -4,3 +4,4 @@ export { default as GeocoderWidget } from './widgets/GeocoderWidget';
 export { default as HistogramWidget } from './widgets/HistogramWidget';
 export { default as PieWidget } from './widgets/PieWidget';
 export { default as LegendWidget } from './widgets/LegendWidget';
+export { default as ScatterPlotWidget } from './widgets/ScatterPlotWidget';