+++ packages/react-workers/src/workerMethods.d.ts
@@ -2,5 +2,6 @@ export enum Methods {
   VIEWPORT_FEATURES = 'viewportFeatures',
   VIEWPORT_FEATURES_FORMULA = 'viewportFeaturesFormula',
   VIEWPORT_FEATURES_HISTOGRAM = 'viewportFeaturesHistogram',
-  VIEWPORT_FEATURES_CATEGORY = 'viewportFeaturesCategory'
+  VIEWPORT_FEATURES_SCATTERPLOT = 'viewportFeaturesScatterPlot',
+  VIEWPORT_FEATURES_CATEGORY = 'viewportFeaturesCategory',
 }