+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Not released
 
+- Fix ScatterplotWidget filters [#153](https://github.com/CartoDB/carto-react/pull/153)
+
+## 1.1.0-alpha.3 (2021-06-17)
+- Add support to CARTO Cloud Native [#142](https://github.com/CartoDB/carto-react/pull/142)
+- Update TypeScript typings to Cloud Native and remove SourceTypes from react-api public api[#147](https://github.com/CartoDB/carto-react/pull/147)
+- Remove SourceTypes from react-api public api [#147](https://github.com/CartoDB/carto-react/pull/147)
+- Add new ScatterplotWidget component [#149](https://github.com/CartoDB/carto-react/pull/149)
+- Update to latest 8.5.0-alpha.10 deck.gl version [#149](https://github.com/CartoDB/carto-react/pull/149)
+- Add support to Cloud Native SQL API [#150](https://github.com/CartoDB/carto-react/pull/150)
+
 ## 1.0.1 (2021-04-12)
 
 - Add basic Typescript typings [#136](https://github.com/CartoDB/carto-react/pull/136)